refactor(access): move function get_permission_required to common viewset
ref: #982 #980
